adding codesplitting in searchMembers.jsx
